+++ src/public/register/register.html
@@ -90,7 +90,7 @@
                         <div class="form-helper-text form-helper-nickname"></div>
                     </div>
                     <div>
-                        <button class="btn form-submit-btn inactivated" type="button" disabled>회원가입</button>
+                        <button class="btn form-submit-btn inactivated" type="button">회원가입</button>
                     </div>
                 </form>
                 <div class="login">
